MonstruosDAO y su controlador al completo
