+++ src/main/java/com/github/Hanselmito/View/Scenes.java
@@ -2,7 +2,7 @@ package com.github.Hanselmito.View;
 
 public enum Scenes {
     ROOT("View/layout.fxml"),
-    MonstruosController("View/MonstruosSaveController.fxml"),
+    MonstruosController("View/MonstruosController.fxml"),
     MenuAdmin("View/MenuAdmin.fxml"),
     Menu("View/Menu.fxml");
