+++ src/main/resources/com/github/Hanselmito/View/MonstruosController.fxml
@@ -3,11 +3,13 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ChoiceBox?>
 <?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane fx:id="anchorPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.github.Hanselmito.View.MonstruosSaveController">
+<AnchorPane fx:id="anchorPane" prefHeight="400.0" prefWidth="701.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.github.Hanselmito.View.MonstruosController">
    <children>
       <TextField fx:id="nombre" layoutX="82.0" layoutY="38.0" promptText="nombre" />
       <TextField fx:id="titulos" layoutX="82.0" layoutY="105.0" promptText="titulos" />
@@ -28,7 +30,15 @@
       <Label layoutX="414.0" layoutY="15.0" text="elementos" />
       <Label layoutX="250.0" layoutY="79.0" text="estados" />
       <Label layoutX="414.0" layoutY="67.0" text="debilidad" />
-      <Button fx:id="insert" layoutX="14.0" layoutY="358.0" mnemonicParsing="false" onMouseClicked="#handleInsertButtonAction" text="Insertar" />
+      <Button fx:id="insert" layoutX="14.0" layoutY="358.0" mnemonicParsing="false" onMouseClicked="#handleInsertButtonAction" text="Insert" />
       <ChoiceBox fx:id="debilidad" layoutX="414.0" layoutY="94.0" prefWidth="150.0" />
+      <TableView fx:id="table" layoutX="607.0" layoutY="38.0" prefHeight="342.0" prefWidth="79.0">
+        <columns>
+          <TableColumn fx:id="ID" prefWidth="78.0" text="ID" />
+        </columns>
+      </TableView>
+      <Button fx:id="update" layoutX="14.0" layoutY="325.0" mnemonicParsing="false" onMouseClicked="#handleUpdateButtonAction" text="Update" />
+      <Button fx:id="Back" layoutX="16.0" layoutY="20.0" mnemonicParsing="false" onMouseClicked="#goBack" text="Back" />
+      <Button fx:id="Delete" layoutX="16.0" layoutY="287.0" mnemonicParsing="false" onMouseClicked="#handleDeleteButtonAction" text="Delete" />
    </children>
 </AnchorPane>
